Revert "QuillのSSRの方法の修正"
